Stock/programs/AI未来予測・戦略策定/projects/AI_Timeline_Forecast/ai_stock_agents/requirements.txt
# AI Stock Agents - Python Dependencies

# LangChain and LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langchain-core>=0.1.0

# LLM and Embeddings
openai>=1.0.0

# Vector Database
chromadb>=0.4.0

# Data Fetching
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0

# API Clients
requests>=2.31.0
python-dotenv>=1.0.0
praw>=7.7.0  # Reddit API
finnhub-python>=2.4.0

# Technical Analysis
pandas-ta>=0.3.14b0
ta-lib>=0.4.0  # Optional - requires C library

# Data Processing
beautifulsoup4>=4.12.0  # SEC EDGAR XML parsing
lxml>=4.9.0  # XML processing for Form 4 filings

# CLI UI
rich>=13.0.0
typer>=0.9.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.65.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type Checking (optional)
mypy>=1.5.0
